requirements.txt
PyYAML~=6.0.2
urllib3~=2.3.0
certifi~=2025.1.31
